Fix the CMake examples build against an installed MOAB
None of the 29 C++ examples could be built from an installed MOAB. Five
independent defects, all pre-existing:

  * examples/CMakeLists.txt read MOAB_C_FLAGS/MOAB_CXX_FLAGS/MOAB_FC_FLAGS,
    but MOABConfig.cmake exports MOAB_CFLAGS/MOAB_CXXFLAGS/MOAB_FCFLAGS, so
    every flag came through empty. It also required CMake 2.8.12, which
    CMake 4 rejects outright, and called enable_language(Fortran) before
    project().

  * MOAB may be configured against MPI via FindMPI rather than the compiler
    wrappers, in which case MOAB_CC/MOAB_CXX are the plain compilers and
    nothing supplied mpi.h. Locate MPI explicitly and add its includes and
    libraries. Third-party headers reachable from MOAB's public headers
    (Eigen via Matrix3.hpp, NetCDF/TempestRemap via the climate headers)
    were missing too; add MOAB_PACKAGE_INCLUDES.

  * MOABConfig.cmake ran separate_arguments() over the exported link line,
    which splits the macOS "-framework Accelerate" pair into two list
    items. Consumers then handed "-framework" and "Accelerate" to
    target_link_libraries() separately and CMake rewrote the latter as
    "-lAccelerate", so every example failed with "ld: framework
    '-lAccelerate' not found". Re-pair "-framework"/"-Xlinker" with the
    argument that follows before de-duplicating.

  * MOABConfig.h has a MOAB_HAVE_MBCOUPLER slot that no build system ever
    populated, so CoupleMGen.cpp always hit its "#error Requires MOAB to be
    built with MBCoupler" even though libmbcoupler was built and installed.
    Set the variable from tools/CMakeLists.txt through the cache so it
    reaches the parent scope where MOABConfig.h is generated, export it as
    MOAB_USE_MBCOUPLER, and gate the example on it. "mbcoupler" is also not
    a target for an external consumer, so resolve it with find_library().

  * TestExodusII.cpp referenced an undeclared test_file_name, and mbex4.cpp
    narrowed unsigned loop counters into a double initializer list.

Verified on macOS arm64: all 29 C++ examples now configure, compile and
link against an installed MOAB; HelloMOAB and TestExodusII run correctly.
The Fortran examples remain gated on ENABLE_IMESH, which the CMake build
does not provide.

+++ examples/beginner/mbex4.cpp
@@ -153,7 +153,7 @@ int main()
         {
             // Get the handle for this vertex:
             moab::EntityHandle handle = scdbox->get_vertex( i, j );
-            double velocity[2]        = { i, j };
+            double velocity[2]        = { static_cast< double >( i ), static_cast< double >( j ) };
 
             // Set the velocity on a vertex:
             MB_CHK_SET_ERR( mbint.tag_set_data( vel_tag, &handle, 1, velocity ), "mbint.tag_set_data(vel_tag) failed" );
